- Rename `basic_shopify_api` to `FastShopifyApi` - Upgrade Shopify API version - Fixed some bugs
