+++ .gitignore
@@ -8,3 +8,4 @@ __pycache__
 /*.egg-info
 /dist
 /build
+.idea/
