+++ LICENSE
@@ -1,6 +1,7 @@
 The MIT License (MIT)
 
-Copyright (c) 2021 Tyler King <tyler@osiset.com>
+Copyright (c) 2021~2022 Tyler King <tyler@osiset.com>
+Copyright (c) 2023 Sokos Lee <SokosLee@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
