--- basic_shopify_api/__version__.py
@@ -1,2 +1,0 @@
-# Current version of libary
-VERSION = "1.0.1"
