+++ fastshopifyapi/__version__.py
@@ -1,0 +1,2 @@
+# Current version of library
+VERSION = "2.0.0"
