+++ tests/fixtures/get_decode_error.json
@@ -1,3 +1,3 @@
-
+{
     "errors": "Not found"
 }
