+++ tests/test_client_graphql.py
@@ -1,6 +1,6 @@
 import pytest
 from .utils import generate_opts_and_sess, local_server_session, async_local_server_session
-from basic_shopify_api import Client, AsyncClient
+from fastshopifyapi import Client, AsyncClient
 
 
 @pytest.mark.usefixtures("local_server")
