+++ tests/test_client_limits.py
@@ -1,7 +1,7 @@
 import pytest
 from http import HTTPStatus
 from .utils import generate_opts_and_sess, local_server_session, async_local_server_session
-from basic_shopify_api import Client, AsyncClient
+from fastshopifyapi import Client, AsyncClient
 
 
 @pytest.mark.usefixtures("local_server")
