+++ tests/utils.py
@@ -1,14 +1,14 @@
 import pytest
 from functools import wraps
 from unittest.mock import patch, PropertyMock
-from basic_shopify_api import Options, Session
-from basic_shopify_api.constants import DEFAULT_MODE
+from fastshopifyapi import Options, Session
+from fastshopifyapi.constants import DEFAULT_MODE
 
 
 def local_server_session(func):
     @wraps(func)
     def wrapper(*args, **kwargs):
-        with patch("basic_shopify_api.Session.base_url", new_callable=PropertyMock) as mock_base_url:
+        with patch("fastshopifyapi.Session.base_url", new_callable=PropertyMock) as mock_base_url:
             mock_base_url.return_value = "http://localhost:8080"
             return func(*args, **kwargs)
     return wrapper
@@ -17,7 +17,7 @@ def local_server_session(func):
 def async_local_server_session(func):
     @wraps(func)
     async def wrapper(*args, **kwargs):
-        with patch("basic_shopify_api.Session.base_url", new_callable=PropertyMock) as mock_base_url:
+        with patch("fastshopifyapi.Session.base_url", new_callable=PropertyMock) as mock_base_url:
             mock_base_url.return_value = "http://localhost:8080"
             return await func(*args, **kwargs)
     return wrapper
